src/main/java/me/raddatz/jwarden/common/service/RequestHelperService.java
package me.raddatz.jwarden.common.service;

import org.springframework.stereotype.Service;
import org.springframework.web.context.request.RequestAttributes;
import org.springframework.web.context.request.RequestContextHolder;
import org.springframework.web.context.request.ServletRequestAttributes;

import javax.servlet.http.HttpServletRequest;

@Service
public class RequestHelperService {

    public String getAuthenticationHeader() {
        var request = getCurrentHttpRequest();

        if (request != null) {
            return request.getHeader("Authentication");
        }
        return "";
    }

    private HttpServletRequest getCurrentHttpRequest() {
        RequestAttributes requestAttributes = RequestContextHolder.getRequestAttributes();
        if (requestAttributes instanceof ServletRequestAttributes) {
            return ((ServletRequestAttributes) requestAttributes).getRequest();
        }
        return null;
    }
}
